+++ fraud-service/src/main/java/com/steve/fraud_service/repository/FraudRepository.java
@@ -4,6 +4,7 @@ import com.steve.fraud_service.model.FraudActivity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -16,5 +17,5 @@ public interface FraudRepository extends JpaRepository<FraudActivity, Long> {
     List<FraudActivity> findByTransactionId(String transactionId);
 
     // Optional: Find all fraud activities with amount greater than a threshold
-    List<FraudActivity> findByAmountGreaterThan(double amount);
+    List<FraudActivity> findByAmountGreaterThan(BigDecimal amount);
 }
